# 4-doc-doc-testing-load-testing/workshop/public/swagger.yml
{
  "swagger": "2.0",
  "info": {
    "description": "TODO API Ynov",
    "version": "1.2.3",
    "title": "Swagger TODO API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3333",
  "tags": [
    {
      "name": "todo",
      "description": "Everything about your Todo",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://localhost:3333"
      }
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/todo": {
      "post": {
        "tags": [
          "todo"
        ],
        "summary": "Ajouter un item todo",
        "description": "",
        "operationId": "addTodo",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "l'objet Todo qui doit être ajouté a la liste",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Todo"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "todo"
        ],
        "summary": "Get all todos",
        "description": "",
        "operationId": "getTodo",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Todo"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/todo/{id}": {
      "patch": {
        "tags": [
          "todo"
        ],
        "summary": "Update specific todo",
        "description": "Returns a single todo",
        "operationId": "patchTodoById",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "ID of todo to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "required": [
        "text"
      ],
      "properties": {
        "text": {
          "type": "string",
          "example": "todo",
          "description": "todo name in the list"
        },
        "done": {
          "type": "boolean",
          "example": false,
          "description": "todo status in the list"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://localhost:3333"
  }
}